Fix warnings

+++ source/core/math/matrix/matrixBase.tpp
@@ -380,7 +380,7 @@ namespace gml
 		// Werte oberhalb der Diagonalen auf 1 bringen
 		for (size_t i = Rows - 1; i > 0; --i)
 		{
-			for (int j = i - 1; j >= 0; --j)
+			for (size_t j = i - 1; j >= 0; --j)
 			{
 				T firstFactor = -mat.at(j, i);
 				T secondFactor = 1;
